add programmatic interface
